+++ window/src/os/x11/connection.rs
@@ -51,6 +51,7 @@ pub struct XConnection {
     pub atom_net_wm_moveresize: Atom,
     pub atom_net_supported: Atom,
     pub atom_net_supporting_wm_check: Atom,
+    pub atom_net_active_window: Atom,
     pub(crate) xrm: RefCell<HashMap<String, String>>,
     pub(crate) windows: RefCell<HashMap<xcb::x::Window, Arc<Mutex<XWindowInner>>>>,
     should_terminate: RefCell<bool>,
@@ -583,6 +584,7 @@ impl XConnection {
         let atom_net_wm_moveresize = Self::intern_atom(&conn, "_NET_WM_MOVERESIZE")?;
         let atom_net_supported = Self::intern_atom(&conn, "_NET_SUPPORTED")?;
         let atom_net_supporting_wm_check = Self::intern_atom(&conn, "_NET_SUPPORTING_WM_CHECK")?;
+        let atom_net_active_window = Self::intern_atom(&conn, "_NET_ACTIVE_WINDOW")?;
 
         let has_randr = conn.active_extensions().any(|e| e == xcb::Extension::RandR);
 
@@ -691,6 +693,7 @@ impl XConnection {
             atom_net_wm_moveresize,
             atom_net_supported,
             atom_net_supporting_wm_check,
+            atom_net_active_window,
             atom_net_wm_icon,
             keyboard,
             kbd_ev,